Simplify the script's logic
